module devide, refactor
