+++ Makefile
@@ -1,16 +1,16 @@
-.PHONY: test run build check 
+.PHONY: test build check echo
+
+echo:
+	@echo good
 
 test :
 	echo "test"
 
-run:
-		@cargo run
+# run:
+# 		@cargo run
 
 build:
 		@cargo build
 
 check:
 		@cargo check
-
-echo:
-	@echo good
