+++ src/main.rs
@@ -1,7 +1,7 @@
-use regex::Regex;
 use skim::prelude::*;
+use std::{io::Cursor, process};
 
-use std::{fs::File, io::Cursor, io::Read, process};
+mod misc;
 
 fn main() {
     let (options, items) = get_params();
@@ -20,6 +20,7 @@ fn main() {
                 .output();
             match output {
                 Ok(output) => {
+                    // TODO: extract as function
                     println!("\n");
                     println!("{}", String::from_utf8_lossy(&output.stdout));
                     println!("{}", String::from_utf8_lossy(&output.stderr));
@@ -41,7 +42,8 @@ fn get_params<'a>() -> (SkimOptions<'a>, Option<Receiver<Arc<dyn SkimItem>>>) {
         .reverse(true)
         .build()
         .unwrap();
-    let commands = match extract_command_from_makefile() {
+    let commands = match misc::extract_command_from_makefile() {
+        // TODO: use some method
         Ok(s) => s,
         Err(e) => {
             print_error(e.to_string());
@@ -54,154 +56,6 @@ fn get_params<'a>() -> (SkimOptions<'a>, Option<Receiver<Arc<dyn SkimItem>>>) {
     (options, Some(items))
 }
 
-/// Makefileからcommandを抽出
-fn extract_command_from_makefile() -> Result<String, &'static str> {
-    let mut file = match File::open("Makefile") {
-        Ok(f) => f,
-        Err(_) => return Err("Makefile not found"),
-    };
-
-    let mut contents = String::new();
-    if file.read_to_string(&mut contents).is_err() {
-        return Err("fail to read Makefile contents");
-    }
-
-    let commands = extract_commands(contents);
-
-    if !commands.is_empty() {
-        Ok(commands.join("\n"))
-    } else {
-        Err("no make command found")
-    }
-}
-
 fn print_error(error_message: String) {
     println!("[ERR] {}", error_message);
-}
-
-fn extract_commands(contents: String) -> Vec<String> {
-    let mut result: Vec<String> = Vec::new();
-    for line in contents.lines() {
-        if let Some(c) = extract_command(line.to_string()) {
-            result.push(c);
-        }
-    }
-
-    result
-}
-
-fn extract_command(line: String) -> Option<String> {
-    let regex = Regex::new(r"^[^.].+:$").unwrap();
-    regex.find(line.as_str()).and_then(|m| {
-        Some(
-            m.as_str()
-                .to_string()
-                .split_once(':')
-                .unwrap()
-                .0
-                .trim()
-                .to_string(),
-        )
-    })
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn extract_commands_test() {
-        struct Case {
-            contents: &'static str,
-            expect: Vec<&'static str>, // NOTE: order of elements of `expect` order should be same as vec function returns
-        }
-        let cases = vec![
-            Case {
-                contents: "\
-    .PHONY: run build check
-
-    run:
-    		@cargo run
-
-    build:
-    		@cargo build
-
-    check:
-    		@cargo check
-
-    echo:
-    	@echo good",
-                expect: vec!["run", "build", "check", "echo"],
-            },
-            Case {
-                contents: "\
-    .PHONY: clone build
-
-    # https://example.com
-    clone:
-    		@git clone https://example.com
-
-    build:
-    		@cargo build",
-                expect: vec!["clone", "build"],
-            },
-        ];
-
-        for case in cases {
-            assert_eq!(case.expect, extract_commands(case.contents.to_string()));
-        }
-    }
-
-    #[test]
-    fn extract_command_test() {
-        struct Case {
-            contents: &'static str,
-            expect: Option<&'static str>,
-        }
-        let cases = vec![
-            Case {
-                contents: "echo:",
-                expect: Some("echo"),
-            },
-            Case {
-                contents: "main.o:",
-                expect: Some("main.o"),
-            },
-            Case {
-                contents: "test::",
-                expect: Some("test"),
-            },
-            Case {
-                contents: "test ::",
-                expect: Some("test"),
-            },
-            Case {
-                contents: "echo",
-                expect: None,
-            },
-            Case {
-                contents: "		@git clone https://example.com",
-                expect: None,
-            },
-            Case {
-                contents: ".PHONY:",
-                expect: None,
-            },
-            Case {
-                contents: ".DEFAULT:",
-                expect: None,
-            },
-            Case {
-                contents: "https://example.com",
-                expect: None,
-            },
-        ];
-
-        for case in cases {
-            assert_eq!(
-                case.expect.map(|e| e.to_string()),
-                extract_command(case.contents.to_string())
-            );
-        }
-    }
 }
